Renamed theme file
