+++ src/EnvironmentMonitor.ReactUi/environment-monitor/src/utilities/baseTheme.ts
